Move userconfig model into a model package
