+++ src/main/java/org/kraaknet/authenticarebankapi/config/model/UserListConfig.java
@@ -1,4 +1,4 @@
-package org.kraaknet.authenticarebankapi.config;
+package org.kraaknet.authenticarebankapi.config.model;
 
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.springframework.boot.context.properties.ConfigurationProperties;
